requirements.txt
Django==5.1.7
djangorestframework==3.16.0
python-decouple==3.8
requests==2.32.3
dj-database-url==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1